make ft_unset function
