+++ Makefile
@@ -6,7 +6,7 @@
 #    By: seojeong <marvin@42.fr>                    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/05/25 19:30:48 by djeon             #+#    #+#              #
-#    Updated: 2021/06/29 22:17:47 by sejpark          ###   ########.fr        #
+#    Updated: 2021/06/30 20:57:55 by sejpark          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -17,11 +17,11 @@ NAME 			= minishell
 
 #READLINE_LIB 	= -lreadline -L/usr/local/opt/readline/lib
 #READLINE_INC	= -I/usr/local/opt/readline/include
-#READLINE_LIB 	= -lreadline -L /Users/$(USER)/.brew/opt/readline/lib
-#READLINE_INC	= -I /Users/$(USER)/.brew/opt/readline/include
+READLINE_LIB 	= -lreadline -L /Users/$(USER)/.brew/opt/readline/lib
+READLINE_INC	= -I /Users/$(USER)/.brew/opt/readline/include
 
-READLINE_LIB	= -lreadline -L/opt/homebrew/opt/readline/lib
-READLINE_INC	= -I/opt/homebrew/opt/readline/include
+#READLINE_LIB	= -lreadline -L/opt/homebrew/opt/readline/lib
+#READLINE_INC	= -I/opt/homebrew/opt/readline/include
 
 LIBFT		= libft.a
 LIBFT_DIR	= libft
@@ -32,7 +32,7 @@ SRC 		= srcs/minishell.c srcs/utils.c srcs/parse.c srcs/exec.c \
 			  srcs/ft_env.c srcs/ft_pwd.c srcs/ft_export.c srcs/export_utils.c \
 			  srcs/signal_handle.c srcs/alloc_token.c srcs/cmd_split.c \
 			  srcs/get_parse_size.c srcs/redir_chk.c srcs/ft_getenv.c \
-			  srcs/redirect.c srcs/redirect_check.c
+			  srcs/redirect.c srcs/redirect_check.c srcs/unset.c
 
 OBJ_DIR 	= objs
 OBJ 		= $(SRC:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
