+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: seojeong <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/14 19:54:23 by djeon             #+#    #+#             */
-/*   Updated: 2021/06/29 21:19:40 by sejpark          ###   ########.fr       */
+/*   Updated: 2021/06/30 20:57:27 by sejpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,33 +19,14 @@
 # include <stdio.h>
 # include <readline/readline.h>
 # include <readline/history.h>
-# include "alloc_token.h"
 # include <fcntl.h>
+# include "structs.h"
+# include "alloc_token.h"
+# include "unset.h"
 
 # define STDIN 			0
 # define STDOUT 		1
 # define STDERR 		2
-
-typedef struct	s_err
-{
-	int				errcode;
-	int				errindex;
-}				t_err;
-
-typedef struct	s_cmd
-{
-	t_token			*cmdline;
-	int				pipe_flag;
-	int				exit_flag;
-	char			quote;
-	char			*(redirect_filename[4]);
-	//index 0 : left redirect char (<, <<);
-	//index 1 : left redirect filename;
-	//index 2 : righ redirect char (>, >>);
-	//index 3 : righ redirect filename;
-	struct s_err	*err_manage;
-	struct s_cmd	*next;
-}				t_cmd;
 
 void				parse(t_cmd **cmd_list, char *line, char **envp);
 t_cmd				*ft_new(char *line, int pipe_flag, char quote, int exit_flag);
