+++ includes/structs.h
@@ -1,0 +1,43 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   structs.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sejpark <sejpark@student.42seoul.kr>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/06/30 17:39:39 by sejpark           #+#    #+#             */
+/*   Updated: 2021/06/30 17:42:40 by sejpark          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef STRUCTS_H
+# define STRUCTS_H
+
+typedef struct	s_err
+{
+	int				errcode;
+	int				errindex;
+}				t_err;
+
+typedef struct	s_token
+{
+	char		*cmd;
+	char		redir_flag;
+}				t_token;
+
+typedef struct	s_cmd
+{
+	t_token			*cmdline;
+	int				pipe_flag;
+	int				exit_flag;
+	char			quote;
+	char			*(redirect_filename[4]);
+	//index 0 : left redirect char (<, <<);
+	//index 1 : left redirect filename;
+	//index 2 : righ redirect char (>, >>);
+	//index 3 : righ redirect filename;
+	struct s_err	*err_manage;
+	struct s_cmd	*next;
+}				t_cmd;
+
+#endif
